Allowing third-party backends to add modules to triton.language.extra
Adding an optional language directory to backends. The contents of the
directory is added to the triton.language.extra namespace during
building.
Updating `triton.language.extra.libdevice` to be able to dispatch to
any backend which adds a libdevice module.
Changing the existing `triton.language.extra.cuda` and
`triton.language.extra.hip` modules to use the new mechanism.